simple-cdd-ext/debian: Disable network configuration when no NIC present
Performing unattented installation using local media (e.g. CD/DVD or
USB stick) on systems without network interface cards (or unrecognized
ones by debian-installer) will trigger netcfg/no_interfaces error
that cannot be preseeded.

The only way to suppress it is to disable network configuration entirely
by setting netcfg/enable preseed to false.

However doing that is tricky as there is no official hook system (except
preseed/run, preseed/early_command, partman/early_command and
preseed/late_command) that can be used to register custom actions on
specific event (e.g. ethtool udeb being installed).

By using recently introduced wrapper functionality that present pre and
post hooks executed before and after wrapped binary execution we can
catch call from anna installer to udpkg with --configure option for
ethdetect package and patch it and it's postinst script to disable
network configuration if ethdetect failed or user chooses "no ethernet
card" in ethdetect/module_select preseed. We cannot patch at post hook
for --unpack option since network install initrd.gz has ethdetect
already unpacked.

This also fixes other non-preseedable errors from choose-mirror and
apt-setup when mirror/* preseeds are given.

+++ debian/early_command.sh
@@ -64,6 +64,58 @@ set_selections_add "$f" 'apt-setup/security_host' "${sec_host}${sec_dir%/}"
 # by corresponding udebs neither loaded nor present in initrd.
 debconf-set-selections "$f"
 
+## Patch postinst scripts for udebs
+
+dpkg_info='/var/lib/dpkg/info'
+
+# Usage: udpkg_pre <exe> ...
+udpkg_pre()
+{
+    # Header
+    echo '#!/bin/sh'
+    echo
+    echo 'set -e'
+    echo
+    echo 'all="$*"'
+    echo
+    echo '# Statically created by pre/post script generator'
+    echo "dpkg_info='$dpkg_info'"
+    echo
+    echo '# Dynamically passed via environment by wrapper'
+    echo '# HOOKS_DIR, PRE, POST, KEEP, THIS'
+    echo
+    echo_function_keep
+    echo
+
+    # Body
+    cat <<'_EOF'
+# ethdetect
+if [ -z "${all##* --configure *ethdetect*}" ]; then
+    # postinst
+    f="$dpkg_info/ethdetect.postinst"
+
+    sed -i "$f" \
+        -e 's,^exec ethdetect$,lsifaces() {\
+    # Based on sed/grep from lsifaces() in ethdetect and netcfg::get_all_ifs()\
+    sed -n -e "/\\W*\\(lo\\|sit[0-9]\\+\\):/d;s/^\\s*\\([a-z0-9]\\+\\):\\s*[0-9]*.\\+$/\\1/p" /proc/net/dev\
+}\
+if ! ethdetect || ! [ -n "$(lsifaces)" ]; then\
+    # Disable network entirely if ethernet card detection fails\
+    . /usr/share/debconf/confmodule\
+    db_set netcfg/enable false\
+fi,'
+
+    keep
+    exit
+fi
+
+# Tell wrapper to keep this hook script
+exit 1
+_EOF
+}
+
+make_wrapper 'udpkg' 'udpkg_pre' '' '1'
+
 ## Cleanup
 
 rm -f "$f" ||:
